Fix handling of lsp-file-watch-ignored-[directories|files] (2735)
When either lsp-file-watch-ignored-directories or
lsp-file-watch-ignored-files are buffer local, the correct value may
not be used when a new directory is added to an already watched
directory. This is due to the watch callback invoking at any time. A
different buffer could be the active buffer when the callback is
invoked causing a different value of either of those two variables to
be used.

This was solved by capturing the value of those two variables at the
time the original watch was created and passing those values into the
watch function.