Capture ignored directory and file list for each project root
The values of `lsp-file-watch-ignored-files` and
`lsp-file-watch-ignored-directories` are now captured on a
per-workspace-root basis. All subsequent handling of watches
underneath that root use the captured value. This means that
customization of those two variables is only honored at the global
level or at a root of an lsp workspace. The documentation of these two
variables have been updated appropriately.

This commit makes a breaking change to the following functions:
* lsp--folder-watch-callback
* lsp-watch-root-folder
* lsp--directory-files-recursively

Each of these now take additional required parameters. The first two
now take both `ignored-files` and `ignored-directories`. The last now
takes `ignored-directories`. Tests have been updated to account for
these API changes.